+++ sandbox/plc4py/plc4py/protocols/modbus/readwrite/ModbusPDUReadFileRecordRequestItem.py
@@ -25,10 +25,10 @@ from dataclasses import dataclass
 
 @dataclass
 class ModbusPDUReadFileRecordRequestItem(PlcMessage):
-            referenceType: short
-            fileNumber: int
-            recordNumber: int
-            recordLength: int
+            referenceType: c_uint8
+            fileNumber: c_uint16
+            recordNumber: c_uint16
+            recordLength: c_uint16
 
 
 
@@ -37,16 +37,16 @@ super().__init__( )
 
 
 
-    def getReferenceType(self) -> short:
+    def getReferenceType(self) -> c_uint8:
         return referenceType
 
-    def getFileNumber(self) -> int:
+    def getFileNumber(self) -> c_uint16:
         return fileNumber
 
-    def getRecordNumber(self) -> int:
+    def getRecordNumber(self) -> c_uint16:
         return recordNumber
 
-    def getRecordLength(self) -> int:
+    def getRecordLength(self) -> c_uint16:
         return recordLength
 
 
@@ -103,13 +103,13 @@ super().__init__( )
         startPos: int = positionAware.getPos()
         curPos: int = 0
 
-                referenceType: short = readSimpleField("referenceType", readUnsignedShort(readBuffer, 8))
+                referenceType: c_uint8 = readSimpleField("referenceType", readUnsignedShort(readBuffer, 8))
 
-                fileNumber: int = readSimpleField("fileNumber", readUnsignedInt(readBuffer, 16))
+                fileNumber: c_uint16 = readSimpleField("fileNumber", readUnsignedInt(readBuffer, 16))
 
-                recordNumber: int = readSimpleField("recordNumber", readUnsignedInt(readBuffer, 16))
+                recordNumber: c_uint16 = readSimpleField("recordNumber", readUnsignedInt(readBuffer, 16))
 
-                recordLength: int = readSimpleField("recordLength", readUnsignedInt(readBuffer, 16))
+                recordLength: c_uint16 = readSimpleField("recordLength", readUnsignedInt(readBuffer, 16))
 
     readBuffer.closeContext("ModbusPDUReadFileRecordRequestItem")
     # Create the instance
@@ -153,4 +153,6 @@ super().__init__( )
             raise RuntimeException(e)
 
         return "\n" + writeBufferBoxBased.getBox().toString()+ "\n"
+
+
